Simplify tileset setup
It's preferable to just set the Tileset.image property to the image to
be loaded. This enables caching fo the image and avoids the  tileset
being seen as an "image collection" tileset where tiles can be
individually added/removed.

I haven't committed the galaxy-tiled.js because it appears to contain
some other code.